Add batch_update to HuaweiSolarBridge.
This function replaces the  method: instead of querying a fixed set of registers,
you can now pass a list of register names that you want to receive.

Context: the DataUpdateCoordinator objects from the Home Assistant integration now keep track
of which entities are enabled (thanks to the  feature that has been added to HA).
This allows us to only request updates for the entities which are currenlty enabled in HA.

**Note**: the license of this project has been changed to AGPL 3.0